Creados DAOs, Services y ServiceImpls de todas las entidades

+++ src/main/java/com/mikelcuenca/app/application/usuario/PermissionRepository.java
@@ -1,9 +1,11 @@
 package com.mikelcuenca.app.application.usuario;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
+public interface PermissionRepository extends JpaRepository<Permission, BigInteger> {
 
 	public Permission getByPermissionName(String permissionName);
